requirements.txt
# Core LangChain dependencies
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# Vector database and embeddings
chromadb>=0.4.0

# AI Provider support
ollama>=0.1.0
openai>=1.0.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# PDF processing
pdfplumber>=0.9.0

# Environment and configuration
python-dotenv>=1.0.0

# Optional: Enhanced features (uncomment if needed)
tiktoken>=0.5.0  # For OpenAI token counting and cost estimation
pydantic>=2.0.0  # For data validation and structured outputs
requests>=2.31.0  # For HTTP requests (usually included with other packages)